Create worker base for workers that use github.
